demo/requirements.txt
prometheus_client>=0.21.0
requests>=2.31.0
python-snappy>=0.7.0
